perf: Refact Service (#2)
