--- src/main/java/dev/patricksilva/controller/Controller.java
@@ -1,13 +1,0 @@
-package dev.patricksilva.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class Controller {
-
-	@GetMapping("/hello")
-	public String hello() {
-		return "Hello world";
-	}
-}
