+++ src/main/java/dev/patricksilva/model/config/WebcamConfig.java
@@ -1,4 +1,4 @@
-package dev.patricksilva.service;
+package dev.patricksilva.model.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -7,13 +7,13 @@ import com.github.sarxos.webcam.Webcam;
 @Configuration
 public class WebcamConfig {
 
-	 /**
+    /**
      * Bean definition for Webcam.
      *
      * @return Webcam bean representing the default webcam configuration.
      */
-	@Bean
-	public Webcam webcam() {
-		return Webcam.getDefault(); // Configuração para a webcam padrão
+    @Bean
+    Webcam webcam() {
+		return Webcam.getDefault();
 	}
 }
