+++ src/main/java/dev/patricksilva/model/entity/Chat.java
@@ -1,4 +1,4 @@
-package UnifyData;
+package dev.patricksilva.model.entity;
 
 public class Chat {
 	public String messages;
