+++ src/main/java/dev/patricksilva/model/entity/Meeting.java
@@ -1,21 +1,19 @@
-package UnifyData;
+package dev.patricksilva.model.entity;
 
 import java.util.Date;
 import java.util.LinkedList;
 
 public class Meeting {
-	
 	private String code;
 	private String title;
 	private Date date;
 	private LinkedList<Client> clientes;
 
 	void createMeeting() {
-		
+
 	}
-	
+
 	void joinMeeting() {
-		
+
 	}
-	
 }
