--- src/main/java/dev/patricksilva/service/WebSocketConfig.java
@@ -1,50 +1,0 @@
-package dev.patricksilva.service;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.task.AsyncTaskExecutor;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-@Configuration
-@EnableWebSocketMessageBroker
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
-
-	/**
-     * Configures the message broker for WebSocket.
-     *
-     * @param config MessageBrokerRegistry to be configured.
-     */
-	@Override
-	public void configureMessageBroker(MessageBrokerRegistry config) {
-		config.enableSimpleBroker("/topic");
-		config.setApplicationDestinationPrefixes("/app");
-	}
-
-	/**
-     * Registers Stomp endpoints.
-     *
-     * @param registry StompEndpointRegistry to register endpoints.
-     */
-	@Override
-	public void registerStompEndpoints(StompEndpointRegistry registry) {
-		registry.addEndpoint("/ws").withSockJS();
-	}
-
-	/**
-     * Bean definition for AsyncTaskExecutor.
-     *
-     * @return AsyncTaskExecutor bean for handling asynchronous tasks.
-     */
-	@Bean(name = "taskExecutor")
-	public AsyncTaskExecutor taskExecutor() {
-		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-		executor.setCorePoolSize(5);
-		executor.setMaxPoolSize(10);
-		executor.setQueueCapacity(25);
-		return executor;
-	}
-}
